[feat] : fin'd 작성 기능 구현 - controller - CRUD 컨트롤러 - 위치 이동(RestController여야 했던게 Controller에 있었음)
